Expanded timeout functionality. User can specify a timeout now. Default, error handling as well.

+++ migrate.sh
@@ -8,6 +8,24 @@ echo
 
 
 
+#If user doesn't provide a timeout, default to 5 seconds. If user provides '0' as the timeout, skip logs questions. Do not allow negative timeouts.  
+if [[ $1 -gt -1 ]]
+then
+	timeout=$1
+else
+	echo "Invalid timeout specified."
+	echo "$(date) - [ERROR] - An invalid timeout was specified. Cannot continue." >> output.log 2>&1
+	echo
+	exit 1
+fi
+if [[ $# -eq 0 ]]
+then
+	timeout=5
+fi
+
+
+
+
 #Ensure that the last job isn't still executing. 
 if [[ -f "migrate.lck" ]]
 then
@@ -51,30 +69,36 @@ do
                 echo && echo
                 while true
                 do
-                        read -t 10 -p "A file named $fileName already exists within ./data/. Would you like to Skip/Overwrite/Rename the file or Exit the program? (s/o/r/e) (Skip if no answer in 10 seconds)" srae
-                        case $srae in
-                                [Ss]* ) echo "$(date) - [WARN] - A file named $fileName already exists within ./data/. This CSV dataset has been skipped." >> output.log 2>&1;
-                                        break;;
+                        read -t $timeout -n 1 -sp "A file named $fileName already exists within ./data/. Would you like to Skip/Overwrite/Rename the file or Exit the program? (s/o/r/e) (Skip after $timeout second timeout)" srae
+			if [! -z $srae ]
+			then
 
-                                [Oo]* ) echo "$(date) - [WARN] - A file named $fileName already exists within ./data/. The old CSV dataset has been overwritten by the new CSV dataset." >> output.log 2>&1;
-                                        gunzip -c $i > data/$fileName 2>&1;
-                                        break;;
+				case $srae in
+					[Ss]* ) echo "$(date) - [WARN] - A file named $fileName already exists within ./data/. This CSV dataset has been skipped." >> output.log 2>&1;
+						break;;
 
-                                [Rr]* ) read -p "Please enter a new file name: " newName;
-                                        while [ -f "data/$newName" ]
-                                        do
-                                                read -p "A file named $newName already exists within ./data/. Please enter a new filename: " newName;
-                                        done;
-                                        echo "$(date) - [WARN] - The file named $fileName already exists within ./data/. The file was renamed to $newName." >> output.log 2>&1;
-                                        gunzip -c $i > data/$newName 2>&1;
-                                        break;;
+					[Oo]* ) echo "$(date) - [WARN] - A file named $fileName already exists within ./data/. The old CSV dataset has been overwritten by the new CSV dataset." >> output.log 2>&1;
+						gunzip -c $i > data/$fileName 2>&1;
+						break;;
 
-                                [Ee]* ) echo "$(date) - [ERROR] - CSV Dataset Migrator was interrupted before completion." >> output.log 2>&1;
-                                        rm migrate.lck;
-                                        exit;;
+					[Rr]* ) read -p "Please enter a new file name: " newName;
+						while [ -f "data/$newName" ]
+						do
+							read -p "A file named $newName already exists within ./data/. Please enter a new filename: " newName;
+						done;
+						echo "$(date) - [WARN] - The file named $fileName already exists within ./data/. The file was renamed to $newName." >> output.log 2>&1;
+						gunzip -c $i > data/$newName 2>&1;
+						break;;
 
-                                * ) echo "Please answer with (Ss/Rr/Aa/Ee)."
-                        esac    
+					[Ee]* ) echo "$(date) - [ERROR] - CSV Dataset Migrator was interrupted before completion." >> output.log 2>&1;
+						rm migrate.lck;
+						exit;;
+
+					* ) echo "Please answer with (Ss/Rr/Aa/Ee)."
+				esac 
+			else
+				break
+			fi 
                 done
         
         #If no files with that name exist... unzip the contents to ./data/ and delete the archive.
@@ -86,14 +110,14 @@ do
                 rm $i
         fi
 done
-echo "$(date) - [INFO] - CSV Dataset Migrator has finished a job. " >> output.log 2>&1
+echo && echo "$(date) - [INFO] - CSV Dataset Migrator has finished a job. " >> output.log 2>&1
 echo "" >> output.log 2>&1
 rm migrate.lck
 
 
 
 
-echo && echo && read -t 5 -n 1 -sp "CSV Dataset Migration has completed. Press 'y' to see the log file. (5s timeout)" input
+echo && echo && read -t $timeout -n 1 -sp "CSV Dataset Migration has completed. Press 'y' to see the log file. ($timeout second timeout)" input
 echo
 
 if [ ! -z $input ]
@@ -104,7 +128,7 @@ fi
 
 
 
-echo && read -t 5 -n 1 -sp "Press 'y' to email the log file to the administrator. (5s timeout)" input
+echo && read -t $timeout -n 1 -sp "Press 'y' to email the log file to the administrator. ($timeout second timeout)" input
 echo
 
 if [ ! -z $input ]
@@ -118,4 +142,3 @@ echo
 
 
 exit
-
